+++ packages/react/src/accordion/accordion-item-context.ts
@@ -1,10 +1,7 @@
+import type { ItemProps, ItemState } from '@zag-js/accordion'
 import { createContext } from '../create-context'
-import { type AccordionContext } from './accordion-context'
 
-export type AccordionItemContext = ReturnType<AccordionContext['getItemState']> & {
-  value: string
-  disabled?: boolean
-}
+export type AccordionItemContext = ItemProps & ItemState
 
 export const [AccordionItemProvider, useAccordionItemContext] = createContext<AccordionItemContext>(
   {