+++ packages/vue/src/accordion/accordion.tsx
@@ -1,11 +1,11 @@
-import { mergeProps } from '@zag-js/vue'
-import { computed, defineComponent } from 'vue'
+import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
+import type { Assign } from '../types'
 import { AccordionProvider } from './accordion-context'
 import { emits, props } from './accordion.props'
 import { useAccordion, type UseAccordionProps } from './use-accordion'
 
-export type AccordionProps = HTMLArkProps<'div'> & UseAccordionProps
+export type AccordionProps = Assign<HTMLArkProps<'div'>, UseAccordionProps>
 
 export const Accordion = defineComponent({
   name: 'Accordion',
@@ -14,8 +14,11 @@ export const Accordion = defineComponent({
   setup(props, { slots, attrs, emit }) {
     const api = useAccordion(props, emit)
     AccordionProvider(api)
-    const mergedProps = computed(() => mergeProps(api.value.rootProps, attrs))
 
-    return () => <ark.div {...mergedProps.value}>{slots?.default?.(api.value)}</ark.div>
+    return () => (
+      <ark.div {...api.value.rootProps} {...attrs}>
+        {slots?.default?.(api.value)}
+      </ark.div>
+    )
   },
 })